requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai>=1.55.3
python-dotenv==1.0.0
pip~=25.2
attrs~=25.3.0
distro~=1.9.0
rsa~=4.9.1
pyasn1~=0.6.1
requests~=2.32.5
zipp~=3.23.0
MarkupSafe~=3.0.2
werkzeug~=3.1.3
click~=8.1.8
jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
aiohttp~=3.9.5
cachetools~=5.5.2
certifi~=2025.8.3
urllib3~=2.5.0
packaging~=24.2
google-api-python-client~=2.183.0
pydantic~=2.11.7
httplib2~=0.31.0
tqdm~=4.67.1
pyparsing~=3.2.5
idna~=3.10
setuptools~=80.9.0
uritemplate~=4.2.0
google-generativeai>=0.8.3
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1